requirements.txt
txtai
openai
sentence_transformers
faiss-cpu